refactor: modularize crate components (#5)
* wip: modularize parts of lib

* refactor: move and imports
