+++ src/config/console.rs
@@ -1,0 +1,5 @@
+pub enum ConsoleConfig {
+    NoReporting,
+    ReportWithoutConsoleColor,
+    ReportWithConsoleColor,
+}
